Rename episodeDetailPage and DetailPodcast and move to component's folder
