+++ src/App.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import './app.css';
 
-import { Route, Routes, useLocation } from 'react-router-dom'
+import { Route, Routes } from 'react-router-dom'
 import MainPage from './pages/MainPage'
-import DetailPage from './pages/DetailPage'
-import EpisodeDetailPage from './pages/EpisodeDetailPage'
+// import EpisodeDetailPage from './pages/EpisodeDetailPage'
 import ErrorPage from './pages/ErrorPage'
 import Header from './components/Header'
 import PodcastPage from './pages/PodcastPage';
@@ -22,27 +21,11 @@ function App () {
         <Route
           path='/podcast/:podcastId'
           element={<PodcastPage />}
-          children={
-            <Route
-              path='/podcast/:podcastId/episode/:episodeId'
-              element={<EpisodeDetailPage />}
-            />
-          }
         />
-        
-        {/* DetailPage and EpisodeDetail could be in one component wiht conditional rendering as they share component */}
-        {/* <Route 
-          path="/podcast/:podcastId" 
-          element={<DetailPage />}
-        /> */}
-        {/* <Route 
-          path="/podcast/:podcastId/episode/:episodeId" 
-          element={<EpisodeDetailPage />}
-        /> */}
-        {/* <Route 
-          path="*" 
-          element={<ErrorPage />}
-        /> */}
+        <Route
+          path='/podcast/:podcastId/episode/:episodeId'
+          element={<PodcastPage />}
+        />
       </Routes>
     </>
   )
