+++ src/components/EpisodeList.jsx
@@ -26,7 +26,7 @@ const EpisodeList = ({ podcastDetail }) => {
     }
     
     return (
-        <div className='p-4 flex flex-col shadow-md rounded-md divide-y-2'>
+        <div className='p-4 divide-y-2'>
             <div className='flex flex-row justify-between p-2'>
                 <h3 className='w-8/12 font-bold'>Title</h3>
                 <h3 className='w-2/12 font-bold'>Date</h3>
